+++ CHANGELOG.md
@@ -2,5 +2,8 @@
 
 This project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0.html).
 
-## [0.0.1] - 2024-10-29
-- Added basic skeletal code for the book
+## [0.1.0] - 2025-08-10
+- Added Chapter 1 - Hello World.
+
+## [0.0.1] - 2025-06-17
+- Added basic skeletal code for the book.